requirements.txt
flask
flake8
pillow
bcrypt
svglib
waitress
cairosvg
requests
reportlab
flask-gzip
python-dotenv
